20171221-Test

+++ iflytransporter-api/src/main/java/com/iflytransporter/api/bean/AuthResp.java
@@ -44,7 +44,7 @@ public class AuthResp{
 
     private String remark;
 
-    private Short status;
+    private Integer status;
 
     private Integer repeatFlag;
 
@@ -230,11 +230,11 @@ public class AuthResp{
 		this.remark = remark;
 	}
 
-	public Short getStatus() {
+	public Integer getStatus() {
 		return status;
 	}
 
-	public void setStatus(Short status) {
+	public void setStatus(Integer status) {
 		this.status = status;
 	}
